% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.Enrichment}
\alias{get.Enrichment}
\title{get.Enrichment}
\usage{
get.Enrichment(
  res,
  db = "GO_Biological_Process_2023",
  ngenes = NULL,
  plot.it = T
)
}
\value{
A result list with - enrichR DF, anrichr barplot
}
\description{
Get enrichR gene function results
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
