% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scSGS}
\alias{scSGS}
\title{scSGS}
\usage{
scSGS(
  mat,
  GoI,
  nHVG = 500,
  HVG_algo = "splinefit",
  show.spline = FALSE,
  calcHVG = FALSE,
  verbose = TRUE,
  nfeatures = 200,
  ncells = 10,
  norm.sep = F,
  rm.mt = FALSE,
  rm.rp = FALSE,
  filter.data = TRUE
)
}
\arguments{
\item{GoI}{A character. Target gene.}

\item{HVG_algo}{A character. Highly variable gene selected method. Default == 'splinefit'. 'seurat' to call Seurat's in-built  HVG selection.}

\item{show.spline}{A Boolean value (TRUE/FALSE), if TRUE, plots spline as a 3D Plotly plot.}

\item{calcHVG}{A Boolean value (TRUE/FALSE), if TRUE, computes highly variable genes, else directly moves ahead to scSGS analysis.}

\item{verbose}{A Boolean value (TRUE/FALSE), if TRUE, displays progress bar.}

\item{nfeatures}{An integer value. Defines the minimum features/genes required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{norm.sep}{A Boolean value (TRUE/FALSE), if TRUE, normalizes KO and WT subset separately.}

\item{rm.mt}{A Boolean value (TRUE/FALSE), if TRUE, removes mitochondrial genes from analysis.}

\item{rm.rp}{A Boolean value (TRUE/FALSE), if TRUE, removes ribosomal protein coding  genes from analysis.}

\item{filter.data}{A Boolean value (TRUE/FALSE), if TRUE, filters cells and genes. Recommended!}

\item{data}{A matrix. Matrix containing scRNA-seq expression with genes as rows and cells as columns.}
}
\value{
A result list with - SGS-DE data.frame, HVG_df data.frame, GoI_Mask List, Known TF list, Highly variable genes List
}
\description{
Predict SGS-responsive genes
}
\examples{
## Load Random Data
counts = matrix(rpois(500*500, lambda = 0.5), ncol=500)
HVG_res <- scSGS(counts, GoI = 10, calcHVG = TRUE)

}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
