% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_splinefit}
\alias{HVG_splinefit}
\title{Spline-HVG}
\usage{
HVG_splinefit(
  data = x,
  degf = 15,
  spar = 0.75,
  nHVGs = 2000,
  show.spline = FALSE,
  use.ndist = T
)
}
\description{
Compute Highly Variable Genes

Spline HVG function
This function computed highly variable genes from an scRNAseq count matrix.
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
