% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_splinefit}
\alias{HVG_splinefit}
\title{HVG_splinefit}
\usage{
HVG_splinefit(
  mat,
  verbose = TRUE,
  dropout.filter = FALSE,
  filter.data = TRUE,
  nfeatures = 200,
  ncells = 10,
  degf = 15,
  spar = 0.75,
  nHVGs = 2000,
  show.spline = FALSE,
  diptest = TRUE,
  use.ndist = TRUE
)
}
\arguments{
\item{verbose}{A Boolean value (TRUE/FALSE), if TRUE, displays progress bar.}

\item{dropout.filter}{A Boolean value (TRUE/FALSE), if TRUE, filters out gene with dropout rate < 0.25 and > 0.75.}

\item{filter.data}{A Boolean value (TRUE/FALSE), if TRUE, filters out gene with dropout rate < 0.25 and > 0.75.}

\item{nfeatures}{An integer value. Defines the minimum features/genes required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{degf}{An integer value. Degrees of freedom for 3D spline computation.}

\item{spar}{A double value. Smoothing parameter for 3D spline computation.}

\item{nHVGs}{An integer value. Number of highly variables to select.}

\item{show.spline}{A Boolean value (TRUE/FALSE), if TRUE, computes Hartigans' dip statistic to test for unimodality.}

\item{use.ndist}{A Boolean value (TRUE/FALSE), if TRUE, uses nearest distance on spline for distance computation, else, uses the genes original location on spline for distance computation.}

\item{adata}{A dgCMatrix object. Sparse matrix contain scRNA-seq expression with genes as rows and cells as columns.}
}
\description{
This function computes highly variable genes from an scRNAseq count matrix.
}
\examples{
## Load Random Data
counts = matrix(rpois(500*500, lambda = 1), ncol=500)
counts = as(counts,'dgCMatrix')
HVG_res <- HVG_splinefit(counts, nHVGs = 100)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
