% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scSGS}
\alias{scSGS}
\title{scSGS}
\usage{
scSGS(
  data,
  GoI,
  nHVG = 500,
  HVG_algo = "splinefit",
  show.spline = FALSE,
  calcHVG = FALSE,
  nfeatures = 200,
  ncells = 10,
  norm_sep = F,
  rm.mt = FALSE,
  rm.rp = FALSE,
  filter_data = TRUE
)
}
\value{
A result list with - DV Genes dataframe, SGS-DE dataframe, TF List
}
\description{
Predict SGS-reponsive genes
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
